commit segunda atividade

+++ leitor_temperatura/src/aplicativo/programa.java
@@ -7,7 +7,7 @@ public class programa {
 	public static void main(String[] args) {
 		Scanner sc=new Scanner(System.in);
 		System.out.println("Digite a temperatura ");
-		int temperatura=sc.nextInt();
+		double temperatura=sc.nextDouble();
 		
 		if(temperatura>30){
 			System.out.println("a temperatura de->"+temperatura);
